// src/services/controlPlaneApi/gen/models/stored-application.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project API
 * Project description API
 *
 * The version of the OpenAPI document: undefined
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { Application } from './application';
// May contain unused imports in some cases
// @ts-ignore
import { ApplicationStatus } from './application-status';

/**
 * 
 * @export
 * @interface StoredApplication
 */
export interface StoredApplication {
    /**
     * 
     * @type {string}
     * @memberof StoredApplication
     */
    'applicationId'?: string;
    /**
     * 
     * @type {Application}
     * @memberof StoredApplication
     */
    'instance'?: Application;
    /**
     * 
     * @type {ApplicationStatus}
     * @memberof StoredApplication
     */
    'status'?: ApplicationStatus;
}