// src/services/controlPlaneApi/gen/models/tenant-configuration.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Project API
 * Project description API
 *
 * The version of the OpenAPI document: undefined
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TenantConfiguration
 */
export interface TenantConfiguration {
    /**
     * 
     * @type {string}
     * @memberof TenantConfiguration
     */
    'name'?: string;
}